# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/game/list": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CronJob Query"
                ],
                "summary": "查詢排程遊戲清單",
                "parameters": [
                    {
                        "type": "string",
                        "description": "group_name",
                        "name": "group_name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"data\":[\"test\"],\"errors\":[]}",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpserver.DataRespSchema"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/api/job/active/{group}/{id}": {
            "put": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CronJob Update"
                ],
                "summary": "啟用排程",
                "parameters": [
                    {
                        "type": "string",
                        "description": "group_name",
                        "name": "group",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "job_id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"errors\":[]}",
                        "schema": {
                            "$ref": "#/definitions/httpserver.SuccessRes"
                        }
                    }
                }
            }
        },
        "/api/job/add": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CronJob Update"
                ],
                "summary": "註冊排程任務",
                "parameters": [
                    {
                        "description": "排程資料",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/cronjob.TaskPayloadReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"errors\":[]}",
                        "schema": {
                            "$ref": "#/definitions/httpserver.SuccessRes"
                        }
                    }
                }
            }
        },
        "/api/job/delete/{group}/{id}": {
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CronJob Update"
                ],
                "summary": "刪除註冊任務",
                "parameters": [
                    {
                        "type": "string",
                        "description": "group_name",
                        "name": "group",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "job_id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"errors\":[]}",
                        "schema": {
                            "$ref": "#/definitions/httpserver.SuccessRes"
                        }
                    }
                }
            }
        },
        "/api/job/game/list": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CronJob Tasks List"
                ],
                "summary": "查詢遊戲排程任務清單 By Game",
                "parameters": [
                    {
                        "type": "string",
                        "description": "group_name",
                        "name": "group_name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "game_type",
                        "name": "game_type",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"data\":[],\"errors\":[]}",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpserver.DataRespSchema"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/cronjob.TaskPayload"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/job/info": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CronJob Query"
                ],
                "summary": "查詢單一排程資料",
                "parameters": [
                    {
                        "type": "string",
                        "description": "group_name",
                        "name": "group_name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "job_id",
                        "name": "job_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"data\":[],\"errors\":[]}",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpserver.DataRespSchema"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/cronjob.TaskPayload"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/job/list": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CronJob Tasks List"
                ],
                "summary": "查詢已註冊排程任務清單 By Group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "group_name",
                        "name": "group_name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"data\":[],\"errors\":[]}",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpserver.DataRespSchema"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/cronjob.TaskPayload"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/job/match/list": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CronJob Tasks List"
                ],
                "summary": "查詢已註冊排程任務清單 By Match",
                "parameters": [
                    {
                        "type": "string",
                        "description": "group_name",
                        "name": "group_name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "match by name",
                        "name": "match",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"data\":[],\"errors\":[]}",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpserver.DataRespSchema"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/cronjob.TaskPayload"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/job/pause/{group}/{id}": {
            "put": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CronJob Update"
                ],
                "summary": "暫停排程",
                "parameters": [
                    {
                        "type": "string",
                        "description": "group_name",
                        "name": "group",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "job_id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"errors\":[]}",
                        "schema": {
                            "$ref": "#/definitions/httpserver.SuccessRes"
                        }
                    }
                }
            }
        },
...
        "/api/job/query/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CronJob Query"
                ],
                "summary": "查詢排程狀態",
                "parameters": [
                    {
                        "type": "string",
                        "description": "job_id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/job/replace": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CronJob Update"
                ],
                "summary": "更新排程任務",
                "parameters": [
                    {
                        "description": "排程資料",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/cronjob.TaskPayloadReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"errors\":[]}",
                        "schema": {
                            "$ref": "#/definitions/httpserver.SuccessRes"
                        }
                    }
                }
            }
        },
        "/api/jobs/delete/{group}": {
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CronJob Update"
                ],
                "summary": "刪除Grop所有註冊任務",
                "parameters": [
                    {
                        "type": "string",
                        "description": "group_name",
                        "name": "group",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"errors\":[]}",
                        "schema": {
                            "$ref": "#/definitions/httpserver.SuccessRes"
                        }
                    }
                }
            }
        },
        "/api/jobs/delete/{group}/{match}": {
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CronJob Update"
                ],
                "summary": "刪除註冊任務 By Match",
                "parameters": [
                    {
                        "type": "string",
                        "description": "group_name",
                        "name": "group",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "match",
                        "name": "match",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"errors\":[]}",
                        "schema": {
                            "$ref": "#/definitions/httpserver.SuccessRes"
                        }
                    }
                }
            }
        },
...
        "/api/jobs/export/{group}": {
            "post": {
                "description": "Export all jobs as a JSON file",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CronJob Import/export"
                ],
                "summary": "匯出任務 By Group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "group_name",
                        "name": "group",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    }
                }
            }
        },
        "/api/jobs/export/{group}/{match}": {
            "post": {
                "description": "Export all jobs as a JSON file",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CronJob Import/export"
                ],
                "summary": "匯出任務 By Match",
                "parameters": [
                    {
                        "type": "string",
                        "description": "group_name",
                        "name": "group",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "match",
                        "name": "match",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "httpserver.DataRespSchema": {
            "type": "object",
            "properties": {
                "data": {},
                "errors": {
                    "type": "array",
                    "items": {}
                }
            }
        },
...
        "httpserver.SuccessRes": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {}
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    }
}